requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.31.0
twilio==8.13.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
